requirements.txt
asgiref==3.2.7
astroid==2.3.3
autopep8==1.5
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
cryptography==2.9.2
defusedxml==0.7.0rc1
Django==3.0.5
django-taggit==1.2.0
django-uuslug==1.2.0
dnspython==1.16.0
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysql-connector-python==8.0.19
mysqlclient==1.4.6
oauthlib==3.1.0
Pillow==7.1.2
protobuf==3.6.1
pycodestyle==2.5.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.4.4
PyMySQL==0.9.3
python-slugify==4.0.0
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
social-auth-app-django==3.4.0
social-auth-core==3.3.3
sorl-thumbnail==12.6.3
sqlparse==0.3.1
text-unidecode==1.3
urllib3==1.25.9
wrapt==1.11.2
